Maths functions -> split into 3 at level 4 for easier readability
