+++ cypher/cypher-docs/src/docs/dev/ql/functions/index.asciidoc
@@ -11,10 +11,9 @@ This section contains information on all functions in the Cypher query language.
 * Scalar functions [<<header-query-functions-scalar, Summary>>|<<query-functions-scalar, Detail>>]
 * Aggregating functions [<<header-query-functions-aggregating, Summary>>|<<query-functions-aggregating, Detail>>]
 * List functions [<<header-query-functions-list, Summary>>|<<query-functions-list, Detail>>]
-* Mathematical functions [<<header-query-functions-mathematical, Summary>>|<<query-functions-mathematical, Detail>>]
-** Numeric functions [<<header-query-functions-numeric, Summary>>|<<query-functions-numeric, Detail>>]
-** Logarithmic functions [<<header-query-functions-logarithmic, Summary>>|<<query-functions-logarithmic, Detail>>]
-** Trigonometric functions [<<header-query-functions-trigonometric, Summary>>|<<query-functions-trigonometric, Detail>>]
+* Mathematical functions - numeric [<<header-query-functions-numeric, Summary>>|<<query-functions-mathematical-numeric, Detail>>]
+* Mathematical functions - logarithmic [<<header-query-functions-logarithmic, Summary>>|<<query-functions-mathematical-logarithmic, Detail>>]
+* Mathematical functions - trigonometric [<<header-query-functions-trigonometric, Summary>>|<<query-functions-mathematical-trigonometric, Detail>>]
 * String functions [<<header-query-functions-string, Summary>>|<<query-functions-string, Detail>>]
 * Spatial functions [<<header-query-functions-spatial, Summary>>|<<query-functions-spatial, Detail>>]
 
@@ -107,14 +106,10 @@ Further details and examples of lists may be found in <<syntax-lists>>.
 |===
 
 
-[[header-query-functions-mathematical]]
-**<<query-functions-mathematical, Mathematical functions>>**
+[[header-query-functions-numeric]]
+**<<query-functions-mathematical-numeric, Mathematical functions - numeric>>**
 
 These functions all operate on numerical expressions only, and will return an error if used on any other values.
-
-
-[[header-query-functions-numeric]]
-**<<query-functions-numeric, Numeric functions>>**
 
 [options="header"]
 |===
@@ -129,7 +124,9 @@ These functions all operate on numerical expressions only, and will return an er
 
 
 [[header-query-functions-logarithmic]]
-**<<query-functions-logarithmic, Logarithmic functions>>**
+**<<query-functions-mathematical-logarithmic, Mathematical functions - logarithmic>>**
+
+These functions all operate on numerical expressions only, and will return an error if used on any other values.
 
 [options="header"]
 |===
@@ -143,7 +140,9 @@ These functions all operate on numerical expressions only, and will return an er
 
 
 [[header-query-functions-trigonometric]]
-**<<query-functions-trigonometric, Trigonometric functions>>**
+**<<query-functions-mathematical-trigonometric, Mathematical functions - trigonometric>>**
+
+These functions all operate on numerical expressions only, and will return an error if used on any other values.
 
 All trigonometric functions operate on radians, unless otherwise specified.
 
@@ -219,9 +218,17 @@ include::query-functions-aggregating.adoc[leveloffset=+1]
 
 include::query-functions-list.adoc[leveloffset=+1]
 
-//Mathematical functions
+//Mathematical functions - numeric
 
-include::query-functions-mathematical.adoc[leveloffset=+1]
+include::query-functions-mathematical-numeric.adoc[leveloffset=+1]
+
+//Mathematical functions - logarithmic
+
+include::query-functions-mathematical-logarithmic.adoc[leveloffset=+1]
+
+//Mathematical functions - trigonometric
+
+include::query-functions-mathematical-trigonometric.adoc[leveloffset=+1]
 
 //String functioms
